started working out np backend
